déplacement de la méthode register dans le SecurityController.php + déplacement du register.html.twig dans un sous-dossier Security
