+++ src/Controller/RegistrationController.php
@@ -25,47 +25,7 @@ class RegistrationController extends AbstractController
         $this->emailVerifier = $emailVerifier;
     }
 
-    #[Route('/register', name: 'app_register')]
-    public function register(Request                     $request,
-                             UserPasswordHasherInterface $userPasswordHasher,
-                             EntityManagerInterface      $entityManager
-    ): Response
-    {
-        $user = new User();
-        $form = $this->createForm(RegistrationFormType::class, $user);
-        $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            // encode the plain password
-            $user->setPassword(
-                $userPasswordHasher->hashPassword(
-                    $user,
-                    $form->get('plainPassword')->getData()
-                )
-            );
-            //attribute ROLE_CONTRIBUTOR to the User
-            $user->setRoles(['ROLE_CONTRIBUTOR']);
-
-            $entityManager->persist($user);
-            $entityManager->flush();
-
-            // generate a signed url and email it to the user
-            $this->emailVerifier->sendEmailConfirmation('app_verify_email', $user,
-                (new TemplatedEmail())
-                    ->from(new Address('admin@wild-series.fr', 'Wild Series'))
-                    ->to($user->getEmail())
-                    ->subject('Veuillez confirmer votre adresse mail')
-                    ->htmlTemplate('registration/confirmation_email.html.twig')
-            );
-
-
-            return $this->redirectToRoute('Home_page');
-        }
-
-        return $this->render('registration/register.html.twig', [
-            'registrationForm' => $form->createView(),
-        ]);
-    }
 
     #[Route('/verify/email', name: 'app_verify_email')]
     public function verifyUserEmail(Request $request, TranslatorInterface $translator): Response
