+++ templates/security/register.html.twig
@@ -1,4 +1,4 @@
-{% extends 'base.html.twig' %}
+{% extends '/base.html.twig' %}
 
 {% block title %}{{ parent() }} - Inscription{% endblock %}
